+++ src/lib/components/asset/SupportedAssetTitle.tsx
@@ -19,14 +19,7 @@ export const SupportedAssetTitle = ({
     !totalSupportedAssetsValue || totalSupportedAssetsValue.eq(0);
 
   return (
-    <Flex
-      w="full"
-      bg="gray.900"
-      py={1}
-      px={4}
-      borderRadius="8px 8px 0px 0px"
-      justifyContent="space-between"
-    >
+    <Flex w="full" py={1} gap={2} align="center" justifyContent="space-between">
       <TableTitle
         title="Supported Assets"
         count={supportedAssets.length}